+++ src/commands/openChangeHead.ts
@@ -1,6 +1,5 @@
 import { SourceControlResourceState, Uri } from "vscode";
 import { Resource } from "../resource";
-import IncomingChangeNode from "../treeView/nodes/incomingChangeNode";
 import { Command } from "./command";
 
 export class OpenChangeHead extends Command {
@@ -9,7 +8,7 @@ export class OpenChangeHead extends Command {
   }
 
   public async execute(
-    arg?: Resource | Uri | IncomingChangeNode,
+    arg?: Resource | Uri | { uri: Uri; type: string },
     ...resourceStates: SourceControlResourceState[]
   ) {
     return this.openChange(arg, "HEAD", resourceStates);