+++ src/commands/openHeadFile.ts
@@ -1,7 +1,6 @@
 import { posix as path } from "path";
 import { commands, Uri, window } from "vscode";
 import { Resource } from "../resource";
-import IncomingChangeNode from "../treeView/nodes/incomingChangeNode";
 import { Command } from "./command";
 
 export class OpenHeadFile extends Command {
@@ -9,15 +8,15 @@ export class OpenHeadFile extends Command {
     super("svn.openHEADFile");
   }
 
-  public async execute(arg?: Resource | Uri | IncomingChangeNode) {
+  public async execute(arg?: Resource | Uri | { uri: Uri; type: string }) {
     let resource: Resource | undefined;
 
     if (arg instanceof Resource) {
       resource = arg;
     } else if (arg instanceof Uri) {
       resource = await this.getSCMResource(arg);
-    } else if (arg instanceof IncomingChangeNode) {
-      resource = new Resource(arg.uri, arg.type, undefined, arg.props, true);
+    } else if (arg && "uri" in arg && "type" in arg) {
+      resource = new Resource(arg.uri, arg.type, undefined, undefined, true);
     } else {
       resource = await this.getSCMResource();
     }