+++ src/historyView/itemLogProvider.ts
@@ -1,4 +1,3 @@
-import * as path from "path";
 import {
   commands,
   Disposable,
@@ -174,17 +173,6 @@ export class ItemLogProvider
       return [];
     }
     if (element === undefined) {
-      const fname = path.basename(this.currentItem.svnTarget.fsPath);
-      const ti = new TreeItem(fname, TreeItemCollapsibleState.Expanded);
-      ti.tooltip = path.dirname(this.currentItem.svnTarget.fsPath);
-      ti.description = path.dirname(this.currentItem.svnTarget.fsPath);
-      ti.iconPath = new ThemeIcon("history");
-      const item = {
-        kind: LogTreeItemKind.TItem,
-        data: ti
-      };
-      return [item];
-    } else {
       const entries = this.currentItem.entries;
       if (entries.length === 0) {
         await fetchMore(this.currentItem);
@@ -208,5 +196,6 @@ export class ItemLogProvider
       }
       return result;
     }
+    return [];
   }
 }